Make AMQP header properties available
Unfortunately, there is no way to pass these headers directly to the
handler function since the handler function is called by the Thrift
generated code and that code has no way to add in extra parameters not
declared in the Thrift file. However, we do know that node is single
thread. So, what we can do is make the properties available as a
property on the AMQPServer object. From there, we can copy them into the
handler's body. Since Node is single threaded, we can be sure that
between the time that this variable is set on the AMQPServer object and
when the handler function initially runs, no other messages will be
processed. As soon as the handler exits, of course, another message may
be received - but sinced we've saved them to a variable in our handler
function, we don't have to worry about that.